app/requirements.txt
psycopg2
pandas
sqlalchemy==1.4
PyYAML
multipledispatch
